refactor: 서버 컴포넌트에서 주소 가져오는 로직 별도 파일로 분리
